Moved cast functions to core.dtypes.cast, tests to test_aggregate
