+++ pandas/core/arrays/__init__.py
@@ -1,8 +1,9 @@
+from pandas.core.dtypes.cast import try_cast_to_ea
+
 from pandas.core.arrays.base import (
     ExtensionArray,
     ExtensionOpsMixin,
     ExtensionScalarOpsMixin,
-    try_cast_to_ea,
 )
 from pandas.core.arrays.boolean import BooleanArray
 from pandas.core.arrays.categorical import Categorical
